feat(rulegen): refactor to visitor pattern
