+++ tasks/rulegen/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "rule_generator"
+name = "rulegen"
 version = "0.1.0"
 edition = "2021"
 publish = false
@@ -12,7 +12,6 @@ handlebars = "4.3.6"
 oxc_allocator = { path = "../../crates/oxc_allocator" }
 oxc_ast = { path = "../../crates/oxc_ast" }
 oxc_parser = { path = "../../crates/oxc_parser" }
-oxc_semantic = { path = "../../crates/oxc_semantic" }
 serde = { workspace = true, features = ["derive"] }
 regex = "1.7.1"
 lazy_static = "1.4.0"
